Update prompt and consolidate workout options into one file
